+++ kaithem/src/docs/changes.md
@@ -4,6 +4,8 @@ Change Log
 ### 0.65.42
 - FreeBoard can now autocomplete tag point names
 - Fix FreeBoard bugs
+- Can now bind native page handlers to subdomains
+- Disallow logging in from any subdomain that contains `__nologin__`, sandboxing those pages down to only what a guest can do.
 
 ### 0.65.41
 - Integrate FreeBoard for no-code dashboard and control interface creation